UpdateGITpracticeVAGubuntuNOTWORKING
Supposed to help practice setting up a git server and using each machine to make changes and then clone or git them from the first machine, i think. I also need practice with git in the blackbox way. this Vagrantfile currently is all ubuntu. ssh 127.0.0.1:2200.
IP addresses of the server is 192.168.6.11.
use vagrant up in the directory of the Vagrantfile created with the vagrant init command.
open that Vagrantfile in any text deal. DELETE everything and paste in the file here. 

currently this does not work