Centos7 git practice setup
fucking guest editions installation issues has forced me to use centos7.
package manager is yum
"33.10, 33.11, 33.12" ip endings.
vagrant ssh centos1 (centos2 or gitsvr)to ssh into the machines from any linux terminal. probably would work in windows command terminal have not tried it, 
for sure works in powershell if you have openssh.

The goal is to and not in this order but somehow, stage a file created with sudo touch test.txt with some text in the file.
then commit i think, or push, or probably one then the other and maybe merge to the branch you must also create the gitserver install in a directory first, cant remember the command. then that directory only is where you create the file. could be the home directory so you might not have to worry about that. 

then from centos2
pull or clone or something the file that was pushed or staged or committed and merged that was done from centos1 and see if you can see the file.
the machines can all connect via ssh using their respective ip addresses. 

yeah.